requirements.txt
python
Flask
classy-json
motor
pytz
idna
discord.py
pymongo
pymongo[srv]
python-dateutil